requirements.txt
langchain>=0.0.267
langchain_ollama>=0.0.1
python-dotenv>=1.0.0
streamlit>=1.24.0
chromadb>=0.4.13
chardet>=5.1.0
langchain_community>=0.0.10
langchain_chroma>=0.0.1
requests>=2.31.0
pymupdf>=1.23.0
JPype1>=1.4.1
sentence-transformers>=2.2.2
langchain-text-splitters>=0.0.1
pdfplumber>=0.9.0
rank-bm25>=0.2.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
beautifulsoup4>=4.12.0